Factor out GFA writing routines
