+++ graph/assemblygraph.cpp
@@ -1585,190 +1585,6 @@ bool AssemblyGraph::saveEntireGraphToFastaOnlyPositiveNodes(const QString& filen
     return true;
 }
 
-static void printTags(QByteArray &out, const std::vector<gfa::tag> &tags) {
-    for (const auto &tag: tags) {
-        out += '\t';
-        out += tag.name[0];
-        out += tag.name[1];
-        out += ":";
-        out += tag.type;
-        out += ":";
-        std::visit([&](const auto& val) {
-            using T = std::decay_t<decltype(val)>;
-            if constexpr (std::is_same_v<T, int64_t>) {
-                out += qPrintable(QString::number(val));
-            } else if constexpr (std::is_same_v<T, float>) {
-                out += qPrintable(QString::number(val));
-            } else if constexpr (std::is_same_v<T, std::string>) {
-                out += val;
-            }
-        }, tag.val);
-    }
-}
-
-QString AssemblyGraph::getGfaSegmentLine(const DeBruijnNode *node, const QString& depthTag) const {
-    QByteArray gfaSequence = node->getSequenceForGfa();
-
-    QByteArray gfaSegmentLine = "S";
-    gfaSegmentLine += "\t" + node->getNameWithoutSign().toLatin1();
-    gfaSegmentLine += "\t" + gfaSequence;
-    gfaSegmentLine += "\tLN:i:" + QString::number(gfaSequence.length()).toLatin1();
-
-    //We use the depthTag to guide how we save the node depth.
-    //If it is empty, that implies that the loaded graph did not have depth
-    //information and so we don't save depth.
-    if (depthTag == "DP")
-        gfaSegmentLine += "\tDP:f:" + QString::number(node->getDepth()).toLatin1();
-    else if (depthTag == "KC" || depthTag == "RC" || depthTag == "FC")
-        gfaSegmentLine += "\t" + depthTag.toLatin1() + ":i:" + QString::number(int(node->getDepth() * gfaSequence.length() + 0.5)).toLatin1();
-
-    //If the user has included custom labels or colours, include those.
-    QString label = getCustomLabel(node);
-    if (!label.isEmpty())
-        gfaSegmentLine += "\tLB:Z:" + label.toLatin1();
-
-    QString rcLabel = getCustomLabel(node->getReverseComplement());
-    if (!rcLabel.isEmpty())
-        gfaSegmentLine += "\tL2:Z:" + rcLabel.toLatin1();
-    if (hasCustomColour(node))
-        gfaSegmentLine += "\tCL:Z:" + getColourName(getCustomColour(node)).toLatin1();
-    if (hasCustomColour(node->getReverseComplement()))
-        gfaSegmentLine += "\tC2:Z:" + getColourName(getCustomColour(node->getReverseComplement())).toLatin1();
-
-    auto tagIt = m_nodeTags.find(node);
-    if (tagIt != m_nodeTags.end())
-        printTags(gfaSegmentLine, tagIt->second);
-
-    return gfaSegmentLine;
-}
-
-QString AssemblyGraph::getGfaLinkLine(const DeBruijnEdge *edge) const {
-    const DeBruijnNode *startingNode = edge->getStartingNode();
-    const DeBruijnNode *endingNode = edge->getEndingNode();
-    bool isJump = edge->getOverlapType() == JUMP;
-
-    QByteArray gfaLinkLine = isJump ? "J\t" : "L\t";
-    gfaLinkLine += qPrintable(startingNode->getNameWithoutSign());
-    gfaLinkLine += '\t';
-    gfaLinkLine += qPrintable(startingNode->getSign());
-    gfaLinkLine += '\t';
-    gfaLinkLine += qPrintable(endingNode->getNameWithoutSign());
-    gfaLinkLine += '\t';
-    gfaLinkLine += qPrintable(endingNode->getSign());
-    gfaLinkLine += '\t';
-    // Emit overlap for normal links and distance for jump links
-    if (isJump) {
-        gfaLinkLine += edge->getOverlap() == 0 ? "*" : qPrintable(QString::number(edge->getOverlap()));
-    } else {
-        gfaLinkLine += qPrintable(QString::number(edge->getOverlap()));
-        gfaLinkLine += "M";
-    }
-
-    if (hasCustomColour(edge)) {
-        gfaLinkLine += "\tCL:Z:";
-        gfaLinkLine += qPrintable(getColourName(getCustomColour(edge)));
-    }
-    if (!edge->isOwnReverseComplement() && hasCustomColour(edge->getReverseComplement())) {
-        gfaLinkLine += "\tC2:Z:";
-        gfaLinkLine += qPrintable(getColourName(getCustomColour(edge->getReverseComplement())));
-    }
-
-    auto tagIt = m_edgeTags.find(edge);
-    if (tagIt != m_edgeTags.end())
-        printTags(gfaLinkLine, tagIt->second);
-
-    return gfaLinkLine;
-}
-
-QString AssemblyGraph::getGfaPathLine(const std::string &name, const Path *path) const {
-    QByteArray gfaPathLine = "P\t";
-    gfaPathLine += name;
-    gfaPathLine += '\t';
-
-    const auto &nodes = path->getNodes();
-    const auto &edges = path->getEdges();
-
-    // edges is one less than nodes for linear paths and of
-    // same length for circular paths
-    for (size_t i = 0; i < edges.size(); ++i) {
-        const auto *edge = edges[i];
-        gfaPathLine += qPrintable(nodes[i]->getName());
-        gfaPathLine += edge->getOverlapType() == JUMP ? ';' : ',';
-    }
-    // Handle last node: for circular paths we're adding extra node here
-    if (nodes.size() == edges.size()) { // circular path
-        gfaPathLine += qPrintable(nodes.front()->getName());
-    } else {
-        gfaPathLine += qPrintable(nodes.back()->getName());
-    }
-
-    return gfaPathLine;
-}
-
-bool AssemblyGraph::saveEntireGraphToGfa(const QString& filename)
-{
-    QFile file(filename);
-    if (!file.open(QIODevice::WriteOnly | QIODevice::Text))
-        return false;
-
-    QTextStream out(&file);
-
-    for (const auto *node : m_deBruijnGraphNodes) {
-        if (node->isPositiveNode())
-            out << getGfaSegmentLine(node, m_depthTag) << '\n';
-    }
-
-    QList<DeBruijnEdge*> edgesToSave;
-    for (auto &entry : m_deBruijnGraphEdges) {
-        DeBruijnEdge * edge = entry.second;
-        if (edge->isPositiveEdge())
-            edgesToSave.push_back(edge);
-    }
-
-    std::sort(edgesToSave.begin(), edgesToSave.end(), DeBruijnEdge::compareEdgePointers);
-
-    for (const auto *edge : edgesToSave)
-        out << getGfaLinkLine(edge) << '\n';
-
-    for (auto it = m_deBruijnGraphPaths.begin(); it != m_deBruijnGraphPaths.end(); ++it)
-        out << getGfaPathLine(it.key(), *it) << '\n';
-
-    return true;
-}
-
-bool AssemblyGraph::saveVisibleGraphToGfa(const QString& filename)
-{
-    QFile file(filename);
-    if (!file.open(QIODevice::WriteOnly | QIODevice::Text))
-        return false;
-
-    QTextStream out(&file);
-
-    for (const auto *node : m_deBruijnGraphNodes) {
-        if (node->thisNodeOrReverseComplementIsDrawn() && node->isPositiveNode())
-            out << getGfaSegmentLine(node, m_depthTag) << '\n';
-    }
-
-    QList<const DeBruijnEdge*> edgesToSave;
-    for (auto &entry : m_deBruijnGraphEdges) {
-        const DeBruijnEdge * edge = entry.second;
-        if (edge->getStartingNode()->thisNodeOrReverseComplementIsDrawn() &&
-            edge->getEndingNode()->thisNodeOrReverseComplementIsDrawn() &&
-            edge->isPositiveEdge())
-            edgesToSave.push_back(edge);
-    }
-
-    std::sort(edgesToSave.begin(), edgesToSave.end(), DeBruijnEdge::compareEdgePointers);
-
-    for (const auto *node : edgesToSave)
-        out << getGfaLinkLine(node) << '\n';
-
-    for (auto it = m_deBruijnGraphPaths.begin(); it != m_deBruijnGraphPaths.end(); ++it)
-        out << getGfaPathLine(it.key(), *it) << '\n';
-
-    return true;
-}
-
 bool AssemblyGraph::hasCustomColour(const DeBruijnNode* node) const {
     auto it = m_nodeColors.find(node);
     return it != m_nodeColors.end() && it->second.isValid();
